Fixed the Visual Studio CMake build. * Added missing CMake function include. * Using ptrdiff_t or int instead of ssize_t. ssize_t is not a standard C++ type. snprintf returns int, not ssize_t. ptrdiff_t fits quite naturally. * Added arguments for the variadic macro parameters. (We should probably remove the variadic parameter from the macro as it's never used) * Avoided preprocessor directives inside macro arguments. * Not linking to LibM when building under MSVC as the math.h implementations are already available in the msvcrt*.dll. (Does torch need the LibM or just some math.h implementation?)

+++ lib/TH/THMemoryFile.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include "THMemoryFile.h"
 #include "THFilePrivate.h"
 
@@ -177,7 +178,7 @@ static int THMemoryFile_mode(const char *mode, int *isReadable, int *isWritable)
       size_t i;                                                           \
       for(i = 0; i < n; i++)                                            \
       {                                                                 \
-        ssize_t nByteWritten;                                           \
+        ptrdiff_t nByteWritten;                                         \
         while (1)                                                       \
         {                                                               \
           ASCII_WRITE_ELEM;                                             \
@@ -479,7 +480,7 @@ static size_t THMemoryFile_writeLong(THFile *self, long *data, size_t n)
     size_t i;
     for(i = 0; i < n; i++)
     {
-      ssize_t nByteWritten;
+      int nByteWritten;
       while (1)
       {
         nByteWritten = snprintf(mfself->storage->data+mfself->position, mfself->storage->size-mfself->position, "%ld", data[i]);
